requirements.txt
# Kafka Client
confluent-kafka==2.3.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# JSON & Configuration
python-dotenv==1.0.0

# Utilities
requests==2.31.0

# Development & Testing
jupyter==1.0.0
notebook==7.0.6
